implement auth on front

+++ client/src/pages/index.tsx
@@ -2,6 +2,7 @@ import { useEffect, useState, useRef } from 'react';
 import { useRouter } from 'next/router';
 import { Table, Thead, Tbody, Tr, Th, Td, Button, Box, Input, Text, Spinner, useDisclosure, AlertDialog, AlertDialogOverlay, AlertDialogContent, AlertDialogHeader, AlertDialogBody, AlertDialogFooter } from '@chakra-ui/react';
 import Layout from '../app/components/Layout';
+import ProtectedRoute from '../../src/app/components/ProtectedRoute';
 
 type Employee = {
   _id: string;
@@ -27,7 +28,17 @@ const Home = () => {
   const fetchEmployees = async () => {
     setLoading(true);
     try {
-      const response = await fetch(`${apiUrl}/api/employees`);
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error('Token is missing');
+      }
+      const response = await fetch(`${apiUrl}/api/employees`, {
+        method: 'GET',
+        headers: {
+          'Authorization': `Bearer ${token}`,
+          'Content-Type': 'application/json'
+        }
+      });
       if (!response.ok) {
         throw new Error('Failed to initiate employee retrieval.');
       }
@@ -39,7 +50,7 @@ const Home = () => {
       setLoading(false);
     }
   };
-
+  
   useEffect(() => {
     fetchEmployees();
   }, [apiUrl]);
@@ -47,9 +58,19 @@ const Home = () => {
   const pollJobStatus = async (jobId: string) => {
     try {
       const interval = setInterval(async () => {
-        const response = await fetch(`${apiUrl}/api/job/${jobId}/status`);
+        const token = localStorage.getItem('token');
+        if (!token) {
+          throw new Error('Token is missing');
+        }
+        const response = await fetch(`${apiUrl}/api/job/${jobId}/status`, {
+          method: 'GET',
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        });
         const data = await response.json();
-
+  
         if (data.state === 'completed') {
           clearInterval(interval);
           if (Array.isArray(data.result)) {
@@ -75,18 +96,23 @@ const Home = () => {
       setLoading(false);
     }
   };
+  
 
   const handleDelete = async () => {
     if (!employeeToDelete) return;
-
     const previousEmployees = employees;
-
-    // Optimistically update the state
     setEmployees(employees.filter(employee => employee._id !== employeeToDelete._id));
-
     try {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        throw new Error('Token is missing');
+      }
       const response = await fetch(`${apiUrl}/api/employees/${employeeToDelete._id}`, {
         method: 'DELETE',
+        headers: {
+          'Authorization': `Bearer ${token}`,
+          'Content-Type': 'application/json'
+        }
       });
       if (!response.ok) {
         throw new Error('Failed to initiate employee deletion.');
@@ -102,6 +128,7 @@ const Home = () => {
       onClose();
     }
   };
+  
 
   const openDeleteConfirm = (employee: Employee) => {
     setEmployeeToDelete(employee);
@@ -113,78 +140,80 @@ const Home = () => {
   );
 
   return (
-    <Layout>
-      {error && <Text color="red.500" mb={4}>{error}</Text>}
-      {loading && <Spinner size="xl" />}
-      <Box mb={4}>
-        <Input
-          placeholder="Search employees..."
-          value={searchTerm}
-          onChange={e => setSearchTerm(e.target.value)}
-          bg="gray.800"
-          color="white"
-        />
-      </Box>
-      <Table variant="simple">
-        <Thead>
-          <Tr>
-            <Th color="white">Name</Th>
-            <Th color="white">Position</Th>
-            <Th color="white">Department</Th>
-            <Th color="white">Date of Hire</Th>
-            <Th color="white">Actions</Th>
-          </Tr>
-        </Thead>
-        <Tbody>
-          {filteredEmployees.map(employee => (
-            <Tr key={employee._id}>
-              <Td>{employee.name}</Td>
-              <Td>{employee.position}</Td>
-              <Td>{employee.department}</Td>
-              <Td>{new Date(employee.dateOfHire).toLocaleDateString('en-US')}</Td>
-              <Td>
-                <Button
-                  colorScheme="blue"
-                  mr={2}
-                  onClick={() => router.push(`/edit-employee?id=${employee._id}`)}
-                >
-                  Edit
+    <ProtectedRoute>
+      <Layout>
+        {error && <Text color="red.500" mb={4}>{error}</Text>}
+        {loading && <Spinner size="xl" />}
+        <Box mb={4}>
+          <Input
+            placeholder="Search employees..."
+            value={searchTerm}
+            onChange={e => setSearchTerm(e.target.value)}
+            bg="gray.800"
+            color="white"
+          />
+        </Box>
+        <Table variant="simple">
+          <Thead>
+            <Tr>
+              <Th color="white">Name</Th>
+              <Th color="white">Position</Th>
+              <Th color="white">Department</Th>
+              <Th color="white">Date of Hire</Th>
+              <Th color="white">Actions</Th>
+            </Tr>
+          </Thead>
+          <Tbody>
+            {filteredEmployees.map(employee => (
+              <Tr key={employee._id}>
+                <Td>{employee.name}</Td>
+                <Td>{employee.position}</Td>
+                <Td>{employee.department}</Td>
+                <Td>{new Date(employee.dateOfHire).toLocaleDateString('en-US')}</Td>
+                <Td>
+                  <Button
+                    colorScheme="blue"
+                    mr={2}
+                    onClick={() => router.push(`/edit-employee?id=${employee._id}`)}
+                  >
+                    Edit
+                  </Button>
+                  <Button colorScheme="red" onClick={() => openDeleteConfirm(employee)}>
+                    Delete
+                  </Button>
+                </Td>
+              </Tr>
+            ))}
+          </Tbody>
+        </Table>
+        <AlertDialog
+          isOpen={isOpen}
+          leastDestructiveRef={cancelRef}
+          onClose={onClose}
+        >
+          <AlertDialogOverlay>
+            <AlertDialogContent>
+              <AlertDialogHeader color={'gray'} fontSize="lg" fontWeight="bold">
+                Delete Employee
+              </AlertDialogHeader>
+
+              <AlertDialogBody color={'gray'}>
+                Are you sure you want to delete this employee? This action cannot be undone.
+              </AlertDialogBody>
+
+              <AlertDialogFooter>
+                <Button ref={cancelRef} onClick={onClose}>
+                  Cancel
                 </Button>
-                <Button colorScheme="red" onClick={() => openDeleteConfirm(employee)}>
+                <Button colorScheme="red" onClick={handleDelete} ml={3}>
                   Delete
                 </Button>
-              </Td>
-            </Tr>
-          ))}
-        </Tbody>
-      </Table>
-      <AlertDialog
-        isOpen={isOpen}
-        leastDestructiveRef={cancelRef}
-        onClose={onClose}
-      >
-        <AlertDialogOverlay>
-          <AlertDialogContent>
-            <AlertDialogHeader color={'gray'} fontSize="lg" fontWeight="bold">
-              Delete Employee
-            </AlertDialogHeader>
-
-            <AlertDialogBody color={'gray'}>
-              Are you sure you want to delete this employee? This action cannot be undone.
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-              <Button ref={cancelRef} onClick={onClose}>
-                Cancel
-              </Button>
-              <Button colorScheme="red" onClick={handleDelete} ml={3}>
-                Delete
-              </Button>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialogOverlay>
-      </AlertDialog>
-    </Layout>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialogOverlay>
+        </AlertDialog>
+      </Layout>
+    </ProtectedRoute>
   );
 };
